+++ apps/transport/lib/transport_web/live/gtfs_diff_select_live.ex
@@ -8,6 +8,11 @@ defmodule TransportWeb.Live.GTFSDiffSelectLive do
   import TransportWeb.Gettext
   alias TransportWeb.GTFSDiffExplain
 
+  import TransportWeb.Live.GTFSDiffSelectLive.Steps
+  import TransportWeb.Live.GTFSDiffSelectLive.Analysis
+  import TransportWeb.Live.GTFSDiffSelectLive.Results
+  import TransportWeb.Live.GTFSDiffSelectLive.Setup
+
   @max_file_size_mb 20
 
   def mount(_params, %{"locale" => locale} = _session, socket) do
@@ -205,14 +210,6 @@ defmodule TransportWeb.Live.GTFSDiffSelectLive do
     Transport.S3.stream_to_s3!(:gtfs_diff, file_path, path, acl: :public_read)
   end
 
-  def uploads_are_valid(%{entries: gtfs}) do
-    gtfs |> Enum.count() == 2 and gtfs |> Enum.all?(&(&1.valid? && &1.done?))
-  end
-
-  defp error_to_string(:too_large), do: "File is too large, must be <#{@max_file_size_mb}MB"
-  defp error_to_string(:too_many_files), do: "You must select 2 files"
-  defp error_to_string(:not_accepted), do: "You have selected an unacceptable file type"
-
   @doc """
   iex> Gettext.put_locale("en")
   iex> translate_target("file", 1)
@@ -248,22 +245,6 @@ defmodule TransportWeb.Live.GTFSDiffSelectLive do
     end)
   end
 
-  def step_completion(current_step, expected_step) do
-    cond do
-      step_progression(current_step) > step_progression(expected_step) -> "done"
-      step_progression(current_step) == step_progression(expected_step) -> "active"
-      true -> ""
-    end
-  end
-
-  defp step_progression(step) do
-    case step do
-      :setup -> 1
-      :analysis -> 2
-      :results -> 3
-    end
-  end
-
   defp drop_empty([]), do: nil
   defp drop_empty(otherwise), do: otherwise
 
@@ -277,383 +258,5 @@ defmodule TransportWeb.Live.GTFSDiffSelectLive do
     |> assign(:error_msg, nil)
     |> assign(:job_running, false)
     |> assign(:selected_file, nil)
-  end
-
-  def steps(%{current_step: _} = assigns) do
-    ~H"""
-    <div id="gtfs-diff-steps" class="container">
-      <ul class="steps-form">
-        <li class={step_completion(@current_step, :setup)}>
-          <div><%= dgettext("validations", "Setup") %></div>
-        </li>
-        <li class={step_completion(@current_step, :analysis)}>
-          <div><%= dgettext("validations", "Analysis") %></div>
-        </li>
-        <li class={step_completion(@current_step, :results)}>
-          <div><%= dgettext("validations", "Results") %></div>
-        </li>
-      </ul>
-    </div>
-    """
-  end
-
-  def preparation_step(%{uploads: _} = assigns) do
-    ~H"""
-    <div id="gtfs-diff-input" class="container" phx-drop-target={@uploads.ref}>
-      <form id="upload-form" phx-submit="gtfs_diff" phx-change="validate">
-        <.upload_drop_zone uploads={@uploads} />
-        <.uploaded_files uploads={@uploads} />
-
-        <button class="button" disabled={not uploads_are_valid(@uploads)} type="submit">
-          <%= dgettext("validations", "Compare") %>
-        </button>
-      </form>
-    </div>
-    """
-  end
-
-  def upload_drop_zone(%{uploads: _} = assigns) do
-    ~H"""
-    <div class="drop-zone panel">
-      <label for={@uploads.ref}>
-        <i class="fa fa-upload" aria-hidden="true"></i>
-        <span>
-          <%= dgettext("validations", "Drop your GTFS files here or click to browse your local drive") %>
-        </span>
-      </label>
-      <.live_file_input upload={@uploads} />
-    </div>
-    """
-  end
-
-  def uploaded_files(%{uploads: _} = assigns) do
-    ~H"""
-    <div id="uploaded-files">
-      <%= for {entry, index} <- upload_entries(@uploads) do %>
-        <.upload uploads={@uploads} entry={entry} index={index} />
-        <%= if index == 0 do %>
-          <.upload_switch uploads={@uploads} />
-        <% end %>
-      <% end %>
-      <.upload_error :for={err <- upload_errors(@uploads)} error={err} />
-    </div>
-    """
-  end
-
-  def upload(%{entry: nil, index: _, uploads: _} = assigns) do
-    ~H"""
-    <article class="upload-entry upload-entry-inactive panel">
-      <h4><%= upload_title(@index) %></h4>
-      <label class="placeholder" for={@uploads.ref}>
-        <%= dgettext("validations", "Please upload some file above") %>
-      </label>
-    </article>
-    """
-  end
-
-  def upload(%{entry: entry, index: _, uploads: uploads} = assigns) do
-    has_errors = length(upload_errors(uploads, entry)) > 0
-
-    classname =
-      if has_errors do
-        "upload-entry upload-entry-errors panel"
-      else
-        "upload-entry panel"
-      end
-
-    assigns =
-      assigns
-      |> assign(:has_errors, has_errors)
-      |> assign(:classname, classname)
-
-    ~H"""
-    <article class={@classname}>
-      <h4><%= upload_title(@index) %></h4>
-      <div class="entry-name">
-        <%= if @entry.valid? do %>
-          <.icon class="fa fa-square-check" title={dgettext("validations", "Valid file")} />
-        <% else %>
-          <.icon class="fa fa-square-xmark" title={dgettext("validations", "Invalid file")} />
-        <% end %>
-        <%= @entry.client_name %>
-      </div>
-      <div class="progress-bar">
-        <progress value={@entry.progress} max="100"><%= @entry.progress %>%</progress>
-        <button
-          type="button"
-          phx-click="cancel-upload"
-          phx-value-ref={@entry.ref}
-          title={dgettext("validations", "Cancel upload or remove file")}
-          aria-label={dgettext("validations", "Cancel upload or remove file")}
-        >
-          <i class="fa fa-xmark"></i>
-        </button>
-      </div>
-      <div :if={@has_errors} class="upload-errors">
-        <%= for err <- upload_errors(@uploads, @entry) do %>
-          <%= error_to_string(err) %>
-        <% end %>
-      </div>
-    </article>
-    """
-  end
-
-  def upload_error(%{error: _} = assigns) do
-    ~H"""
-    <p class="alert alert-danger"><i class="fa fa-square-xmark"></i> <%= error_to_string(@error) %></p>
-    """
-  end
-
-  defp upload_title(index) do
-    if index == 0 do
-      dgettext("validations", "Reference GTFS")
-    else
-      dgettext("validations", "Modified GTFS")
-    end
-  end
-
-  def upload_switch(%{uploads: _} = assigns) do
-    ~H"""
-    <div>
-      <button
-        disabled={length(@uploads.entries) != 2}
-        class="button-outline primary small"
-        type="button"
-        title={dgettext("validations", "Switch files")}
-        aria-label={dgettext("validations", "Switch files")}
-        phx-click="switch-uploads"
-      >
-        <i class="fa fa-arrow-right-arrow-left"></i>
-      </button>
-    </div>
-    """
-  end
-
-  defp icon(%{title: _, class: _} = assigns) do
-    ~H"""
-    <i class={@class} title={@title} aria-label={@title}></i>
-    """
-  end
-
-  def upload_entries(uploads) do
-    uploads.entries
-    |> Enum.concat([nil, nil])
-    |> Enum.take(2)
-    |> Enum.with_index()
-  end
-
-  def diff_natures do
-    [
-      {"add", dgettext("validations", "added"), "green"},
-      {"update", dgettext("validations", "updated"), "orange"},
-      {"delete", dgettext("validations", "deleted"), "red"}
-    ]
-  end
-
-  def diff_summaries_for_file(%{selected_file: _, diff_summary: _} = assigns) do
-    ~H"""
-    <ul>
-      <.diff_summary_for_file
-        :for={{nature, translation, css_class} <- diff_natures()}
-        summary={@diff_summary[nature]}
-        translation={translation}
-        selected_file={@selected_file}
-        class={css_class}
-      />
-    </ul>
-    """
-  end
-
-  def diff_summary_for_file(%{summary: _, selected_file: _, translation: _, class: _} = assigns) do
-    ~H"""
-    <div :if={@summary}>
-      <%= for {{file, _nature, target}, n} <- @summary do %>
-        <li :if={file == @selected_file}>
-          <span class={@class}><%= @translation %> &nbsp;</span><%= translate_target(target, n) %>
-        </li>
-      <% end %>
-    </div>
-    """
-  end
-
-  def diff_summaries(%{files_with_changes: _, selected_file: _, diff_summary: _, diff_explanations: _} = assigns) do
-    ~H"""
-    <div class="pt-24">
-      <div class="dashboard">
-        <.navigation files_with_changes={@files_with_changes} selected_file={@selected_file} />
-        <.differencies diff_summary={@diff_summary} selected_file={@selected_file} diff_explanations={@diff_explanations} />
-      </div>
-    </div>
-    """
-  end
-
-  def navigation(%{files_with_changes: _, selected_file: _} = assigns) do
-    ~H"""
-    <aside class="side-menu" role="navigation">
-      <ul>
-        <.select_file_navigation_link :for={file <- @files_with_changes} file={file} selected_file={@selected_file} />
-      </ul>
-    </aside>
-    """
-  end
-
-  def select_file_navigation_link(%{file: _, selected_file: _} = assigns) do
-    assigns =
-      assigns
-      |> assign(
-        :class,
-        if assigns[:file] == assigns[:selected_file] do
-          "active"
-        end
-      )
-
-    ~H"""
-    <li>
-      <a class={@class} phx-click="select-file" phx-value-file={@file}>
-        <code><%= @file %></code>
-      </a>
-    </li>
-    """
-  end
-
-  def differencies(%{diff_summary: _, selected_file: _, diff_explanations: _} = assigns) do
-    ~H"""
-    <div class="main">
-      <p><%= dgettext("validations", "Differences Overview") %></p>
-      <.diff_summaries_for_file diff_summary={@diff_summary} selected_file={@selected_file} />
-      <%= if assigns[:diff_explanations] do %>
-        <% active_explanations =
-          @diff_explanations
-          |> Enum.filter(fn {file, _} -> file == @selected_file end)
-          |> Enum.map(fn {_, explanation} -> explanation end) %>
-        <p :if={not Enum.empty?(active_explanations)}><%= dgettext("validations", "Detail") %></p>
-        <ul>
-          <%= for explanation <- active_explanations do %>
-            <li>
-              <%= explanation %>
-            </li>
-          <% end %>
-        </ul>
-      <% end %>
-    </div>
-    """
-  end
-
-  def results_step(
-        %{
-          diff_explanations: _,
-          diff_file_url: _,
-          diff_summary: _,
-          error_msg: _,
-          files_with_changes: _,
-          gtfs_original_file_name_1: _,
-          gtfs_original_file_name_2: _,
-          selected_file: _
-        } = assigns
-      ) do
-    ~H"""
-    <div class="container gtfs-diff-results">
-      <div :if={@diff_file_url} class="panel">
-        <h4>
-          <%= dgettext("validations", "GTFS Diff is available for") %>
-          <%= link(dgettext("validations", "download"),
-            to: @diff_file_url,
-            target: "_blank"
-          ) %>
-        </h4>
-        <%= raw(
-          dgettext(
-            "validations",
-            "<a href=\"%{spec}\">Read</a> the GTFS Diff specification to understand how differences between GTFS are expressed",
-            spec: "https://github.com/etalab/gtfs_diff/blob/main/specification.md"
-          )
-        ) %>.
-        <%= if @diff_summary do %>
-          <div class="pt-24">
-            <%= if @diff_summary == %{} do %>
-              <%= similar_files(@gtfs_original_file_name_1, @gtfs_original_file_name_2) %>
-            <% else %>
-              <%= different_files(@gtfs_original_file_name_1, @gtfs_original_file_name_2) %>
-              <.diff_summaries
-                diff_explanations={@diff_explanations}
-                diff_summary={@diff_summary}
-                files_with_changes={@files_with_changes}
-                selected_file={@selected_file}
-              />
-            <% end %>
-          </div>
-        <% else %>
-          <%= if @error_msg do %>
-            <.validation_error error_msg={@error_msg} />
-          <% else %>
-            <div class="pt-24">
-              <%= dgettext("validations", "analyzing found differences...") %>
-            </div>
-          <% end %>
-        <% end %>
-      </div>
-
-      <button class="button primary" type="button" phx-click="start-over">
-        <i class="fa fa-rotate-left"></i>&nbsp;<%= dgettext("validations", "Start over") %>
-      </button>
-    </div>
-    """
-  end
-
-  def similar_files(file1, file2) do
-    dgettext(
-      "validations",
-      "The GTFS files <code>%{gtfs_original_file_name_2}</code> and <code>%{gtfs_original_file_name_1}</code> are similar.",
-      gtfs_original_file_name_1: file1,
-      gtfs_original_file_name_2: file2
-    )
-    |> raw()
-  end
-
-  def different_files(file1, file2) do
-    dgettext(
-      "validations",
-      "The GTFS file <code>%{gtfs_original_file_name_2}</code> has differences with the GTFS file <code>%{gtfs_original_file_name_1}</code>, as summarized below:",
-      gtfs_original_file_name_1: file1,
-      gtfs_original_file_name_2: file2
-    )
-    |> raw()
-  end
-
-  def validation_error(%{error_msg: _} = assigns) do
-    ~H"""
-    <div class="pt-24">
-      <%= dgettext(
-        "validations",
-        "An error occurred while interpreting the results. Note that the report is still available as download. Error:"
-      ) %>
-      <.error_message error_msg={@error_msg} />
-    </div>
-    """
-  end
-
-  def error_message(%{error_msg: _} = assigns) do
-    ~H"""
-    <span class="red"><%= @error_msg %></span>
-    """
-  end
-
-  def analysis_step(%{diff_logs: _, error_msg: _} = assigns) do
-    ~H"""
-    <div class="container">
-      <div class="panel">
-        <h4><%= dgettext("validations", "Processing") %></h4>
-        <%= for log <- Enum.reverse(@diff_logs) do %>
-          <div>
-            <%= raw(log) %>...
-          </div>
-        <% end %>
-      </div>
-
-      <div :if={@error_msg}>
-        <.error_message error_msg={@error_msg} />
-      </div>
-    </div>
-    """
   end
 end
